Begin redesign for modularity
